+++ app/Http/Requests/StoreRecipeRequest.php
@@ -24,8 +24,32 @@ class StoreRecipeRequest extends FormRequest
     {
         return [
             'title' => 'required|max:255|unique:recipes,title',
-            'product_ids' => 'required|array|min:1',
-            'product_ids.*' => 'integer|exists:products,id',
+            'products' => 'required|array|min:1',
+            'products.*.product_id' => 'required|integer|exists:products,id',
+            'products.*.quantity' => 'required|integer|min:1',
         ];
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function messages(): array
+    {
+        return [
+            'products.*.quantity.required' => 'Введите количество',
+            'products.*.quantity.integer' => 'Количество должно быть целым числом',
+            'products.*.quantity.min' => 'Количество должно быть не менее :min',
+            'products.required' => 'Необходимо добавить хотя бы один продукт',
+        ];
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    protected function prepareForValidation(): void
+    {
+        $this->merge([
+            'products' => json_decode($this->request->get('products'), true),
+        ]);
     }
 }