+++ app/Models/Recipe.php
@@ -27,7 +27,8 @@ class Recipe extends Model
      */
     public function products(): BelongsToMany
     {
-        return $this->belongsToMany(Product::class);
+        return $this->belongsToMany(Product::class)
+            ->withPivot('quantity');
     }
 
     /**